move validator to utils/
