+++ client/containers/Pub/SpubHeader/SpubHeader.tsx
@@ -1,10 +1,11 @@
 import React, { useMemo, useRef, useState } from 'react';
 
 import { DocJson, Pub, PubPageData } from 'types';
+import { PendingChangesProvider } from 'components';
 import { apiFetch } from 'client/utils/apiFetch';
 import { useSticky } from 'client/utils/useSticky';
-import { PendingChangesProvider } from 'components';
 import { usePendingChanges } from 'utils/hooks';
+import { createSubmissionValidator, validateSubmission } from 'utils/submission/validate';
 
 import { usePubContext } from '../pubHooks';
 import PubHeader from '../PubHeader';
@@ -14,7 +15,6 @@ import InstructionsTab from './InstructionsTab';
 import DetailsTab from './DetailsTab';
 import ContributorsTab from './ContributorsTab';
 import SpubHeaderToolBar from './SpubHeaderToolBar';
-import { createSubmissionValidator, validateSubmission } from './validate';
 
 require('./spubHeader.scss');
